Move all Delve classes to same folder and namespace
