+++ GameServer/packets/Client/168/DetailDisplayHandler.cs
@@ -17,12 +17,11 @@
  *
  */
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using DOL.Database;
+using DOL.GS.Delve;
 using DOL.GS.Effects;
 using DOL.GS.Quests;
 using DOL.GS.RealmAbilities;
